requirements.txt
asgiref==3.2.10
coverage==5.2.1
Django==3.0.8
django-cors-headers==3.4.0
django-environ==0.4.5
django-filter==2.3.0
djangorestframework==3.11.0
flake8==3.8.3
mccabe==0.6.1
model-mommy==2.0.0
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1
